+++ Domain/Entities/CoursesAndGroups/Lecture.cs
@@ -11,7 +11,7 @@ namespace ClassJournals.Domain.Entities.CoursesAndGroups
                                            // щоб можна було розділити теми по лекціямм (в ідеалі щоб, коли
                                            // позначав лекцію в розкладі, то воно автоматично розприділяло теми
                                            // лекцій по днях)?
-        public double Hours { get; set; }
+        public long Hours { get; set; }
         public int Raiting { get; set; }
 
         public int CurrentLectorId { get; set; }
